Reorganize code to modules
