+++ src/Changelogged/Common.hs
@@ -1,0 +1,7 @@
+module Changelogged.Common
+  ( module Changelogged.Common.Utils
+  , module Changelogged.Common.Types
+  ) where
+
+import Changelogged.Common.Utils
+import Changelogged.Common.Types
