+++ src/Changelogged/Common/Utils.hs
@@ -1,0 +1,7 @@
+module Changelogged.Common.Utils
+  ( module Changelogged.Common.Utils.Pure
+  , module Changelogged.Common.Utils.IO
+  ) where
+
+import Changelogged.Common.Utils.Pure
+import Changelogged.Common.Utils.IO
