+++ src/Changelogged/Config.hs
@@ -1,8 +1,5 @@
-{-# LANGUAGE DeriveGeneric #-}
-{-# LANGUAGE TemplateHaskell #-}
 module Changelogged.Config where
 
-import Data.Aeson.TH
 import Data.Either.Combinators (rightToMaybe)
 import Data.Monoid ((<>))
 import Data.Text (Text)
@@ -10,43 +7,9 @@ import qualified Data.Text as Text
 import qualified Data.Yaml as Yaml
 
 import qualified Turtle
-import GHC.Generics
 
-import Changelogged.Types ()
-import Changelogged.Utils ()
-import Changelogged.Pure
-
-data Config = Config
-  { configChangelogs    :: [ChangelogConfig]
-  , configBranch        :: Maybe Text
-  } deriving (Eq, Show, Generic)
-
-data LevelHeaders = LevelHeaders
-  { levelHeadersApp   :: Maybe Text
-  , levelHeadersMajor :: Maybe Text
-  , levelHeadersMinor :: Maybe Text
-  , levelHeadersFix   :: Maybe Text
-  , levelHeadersDoc   :: Maybe Text
-  } deriving (Eq, Show, Generic)
-
-data ChangelogConfig = ChangelogConfig
-  { changelogChangelog     :: Turtle.FilePath
-  , changelogLevelHeaders  :: LevelHeaders
-  , changelogWatchFiles    :: Maybe [Turtle.FilePath]
-  , changelogIgnoreFiles   :: Maybe [Turtle.FilePath]
-  , changelogIgnoreCommits :: Maybe [Text]
-  , changelogVersionFiles  :: Maybe [VersionFile]
-  } deriving (Eq, Show, Generic)
-
-data VersionPattern = VersionPattern
-  { versionPatternVariable  :: Text
-  , versionPatternSeparator :: Text
-  } deriving (Show, Eq, Generic)
-
-data VersionFile = VersionFile
-  { versionFilePath :: Turtle.FilePath
-  , versionFileVersionPattern :: VersionPattern
-  } deriving (Show, Eq, Generic)
+import Changelogged.Common
+import Changelogged.Aeson ()
 
 defaultLevelHeaders :: LevelHeaders
 defaultLevelHeaders = LevelHeaders
@@ -88,9 +51,3 @@ ppConfig Config{..} = mconcat
 
     _    ?: Nothing = ""
     name ?: Just val = name !: val
-
-deriveJSON (jsonDerivingModifier "VersionPattern") ''VersionPattern
-deriveJSON (jsonDerivingModifier "VersionFile") ''VersionFile
-deriveJSON (jsonDerivingModifier "LevelHeaders") ''LevelHeaders
-deriveJSON (jsonDerivingModifier "Changelog") ''ChangelogConfig
-deriveJSON (jsonDerivingModifier "Config") ''Config
